+++ src/App/Fossa/PreflightChecks.hs
@@ -7,19 +7,23 @@ module App.Fossa.PreflightChecks (
   guardWithPreflightChecks,
 ) where
 
-import App.Docs (apiKeyUrl)
+import App.Docs (apiKeyUrl, apiTokenDocsUrl, fossaConfigDocsUrl, rolesDocsUrl)
+import App.Support (reportDefectMsg)
 import App.Types (ProjectMetadata, ProjectRevision)
 import Control.Carrier.Debug (ignoreDebug)
-import Control.Carrier.Diagnostics (Diagnostics, errCtx)
+import Control.Carrier.Diagnostics (Diagnostics, errCtx, errDoc, errHelp)
 import Control.Carrier.FossaApiClient (runFossaApiClient)
 import Control.Carrier.Stack (context)
-import Control.Effect.Diagnostics (ToDiagnostic, fatalOnIOException, fatalText)
+import Control.Effect.Diagnostics (ToDiagnostic, errSupport, fatal, fatalOnIOException)
 import Control.Effect.FossaApiClient (FossaApiClient, getCustomBuildPermissions, getOrganization, getSubscription, getTokenType)
 import Control.Effect.Lift (Has, Lift, sendIO)
 import Control.Monad (void, when)
+import Data.Error (createErrataWithHeaderOnly)
+import Data.Text
 import Data.Text.IO qualified as TIO
 import Diag.Diagnostic (ToDiagnostic (..))
-import Effect.Logger (Logger, logDebug, pretty, vsep)
+import Effect.Logger (renderIt)
+import Errata (Errata (..))
 import Fossa.API.Types (ApiOpts, CustomBuildUploadPermissions (..), Organization (orgSupportsPreflightChecks), ProjectPermissionStatus (..), ReleaseGroupPermissionStatus (..), Subscription (..), SubscriptionResponse (..), TokenType (..), TokenTypeResponse (..))
 import Path (
   File,
@@ -30,20 +34,16 @@ import Path (
   (</>),
  )
 import Path.IO (getTempDir, removeFile)
-import Text.Pretty.Simple (pShow)
 
 data PreflightCommandChecks
   = AnalyzeChecks ProjectRevision ProjectMetadata
   | TestChecks
   | ReportChecks
-  | ConatinerAnalzyeChecks
-  | ContainerTestChecks
   | AssertUserDefinedBinariesChecks
 
 guardWithPreflightChecks ::
   ( Has Diagnostics sig m
   , Has (Lift IO) sig m
-  , Has Logger sig m
   ) =>
   ApiOpts ->
   PreflightCommandChecks ->
@@ -54,7 +54,6 @@ preflightChecks ::
   ( Has Diagnostics sig m
   , Has (Lift IO) sig m
   , Has FossaApiClient sig m
-  , Has Logger sig m
   ) =>
   PreflightCommandChecks ->
   m ()
@@ -65,26 +64,24 @@ preflightChecks cmd = context "preflight-checks" $ do
   sendIO $ removeFile (tmpDir </> preflightCheckFileName)
 
   -- Check for valid API Key and if user can connect to fossa app
-
-  org <- errCtx InvalidApiKeyErr getOrganization
+  org <- errHelp InvalidApiKeyErr $ errDoc apiKeyUrl getOrganization
 
   when (orgSupportsPreflightChecks org) $
     void $
-      case cmd of
-        TestChecks -> do
-          tokenType <- getTokenType
-          fullAccessTokenCheck tokenType
-        AnalyzeChecks rev metadata -> do
-          customBuildPermissions <- getCustomBuildPermissions rev metadata
-          logDebug $ "This is the permission data" <> pretty (pShow (customBuildPermissions))
-          uploadBuildPermissionsCheck customBuildPermissions
-        _ -> do
-          tokenType <- getTokenType
-          subscription <- getSubscription
-          fullAccessTokenCheck tokenType
-          premiumSubscriptionCheck subscription
-
-  void $ errCtx InvalidApiKeyErr getOrganization
+      errSupport (renderIt reportDefectMsg) $
+        case cmd of
+          TestChecks -> do
+            tokenType <- getTokenType
+            fullAccessTokenCheck tokenType
+          AnalyzeChecks rev metadata -> do
+            customBuildPermissions <- getCustomBuildPermissions rev metadata
+            uploadBuildPermissionsCheck customBuildPermissions
+          ReportChecks -> do
+            tokenType <- getTokenType
+            subscription <- getSubscription
+            fullAccessTokenCheck tokenType
+            premiumSubscriptionCheck subscription
+          _ -> pure ()
 
 uploadBuildPermissionsCheck :: Has Diagnostics sig m => CustomBuildUploadPermissions -> m ()
 uploadBuildPermissionsCheck CustomBuildUploadPermissions{..} =
@@ -92,24 +89,52 @@ uploadBuildPermissionsCheck CustomBuildUploadPermissions{..} =
     Just releaseGroupPermissionStatus ->
       case releaseGroupPermissionStatus of
         ValidReleaseGroupPermission -> pure ()
-        InvalidEditReleaseGroupPermission -> fatalText "You do not have permission to edit this release group"
-        InvalidCreateTeamProjectsForReleaseGroupPermission -> fatalText "You do not have permission to add projects to all teams which include this release group"
+        InvalidEditReleaseGroupPermission ->
+          errDoc rolesDocsUrl
+            . errHelp permissionHelpMsg
+            $ fatal EditReleaseGroupPermissionErr
+        InvalidCreateTeamProjectsForReleaseGroupPermission ->
+          errDoc rolesDocsUrl
+            . errHelp permissionHelpMsg
+            $ fatal CreateTeamProjectsForReleaseGroupPermissionErr
     Nothing ->
       case projectPermissionStatus of
         ValidProjectPermission -> pure ()
-        InvalidEditProjectPermission -> fatalText "You do not have permission to edit this project"
-        InvalidCreateProjectPermission -> fatalText "You do not have permission to create project"
-        InvalidCreateTeamProjectPermission -> fatalText "You do not have permission to a project for the specified team"
-        InvalidCreateProjectOnlyToTeamPermission -> fatalText "You permissions create projects but only for a specified team"
+        InvalidEditProjectPermission ->
+          errDoc rolesDocsUrl
+            . errHelp permissionHelpMsg
+            $ fatal EditProjectPermissionErr
+        InvalidCreateProjectPermission ->
+          errDoc rolesDocsUrl
+            . errHelp permissionHelpMsg
+            $ fatal CreateProjectPermissionErr
+        InvalidCreateTeamProjectPermission ->
+          errDoc rolesDocsUrl
+            . errHelp permissionHelpMsg
+            $ fatal CreateTeamProjectPermissionErr
+        InvalidCreateProjectOnlyToTeamPermission ->
+          errDoc fossaConfigDocsUrl
+            . errHelp ("Ensure that you have specified a team to add this project to" :: Text)
+            $ fatal CreateProjectOnlyToTeamPermissionErr
+  where
+    permissionHelpMsg :: Text
+    permissionHelpMsg = "Contact your FOSSA organization admin to grant you proper permissions"
 
 fullAccessTokenCheck :: Has Diagnostics sig m => TokenTypeResponse -> m ()
 fullAccessTokenCheck TokenTypeResponse{..} = case tokenType of
-  Push -> fatalText "You are using a push-only token when you need a full-access token"
+  Push ->
+    errHelp ("Ensure you are using a `Full Access` API token" :: Text)
+      . errDoc apiTokenDocsUrl
+      . errCtx ("You are currently using a `Push Only` API token" :: Text)
+      $ fatal TokenTypeErr
   _ -> pure ()
 
 premiumSubscriptionCheck :: Has Diagnostics sig m => SubscriptionResponse -> m ()
 premiumSubscriptionCheck SubscriptionResponse{..} = case subscription of
-  Free -> fatalText "You have free subscription when you need premium subscription"
+  Free ->
+    errHelp ("To proceed, please upgrade your subscription" :: Text)
+      . errCtx ("You currently have a free subscription" :: Text)
+      $ fatal SubscriptionTypeErr
   _ -> pure ()
 
 preflightCheckFileName :: Path Rel File
@@ -117,8 +142,52 @@ preflightCheckFileName = $(mkRelFile "preflight-check.txt")
 
 data InvalidApiKeyErr = InvalidApiKeyErr
 instance ToDiagnostic InvalidApiKeyErr where
+  renderDiagnostic :: InvalidApiKeyErr -> Errata
   renderDiagnostic InvalidApiKeyErr =
-    vsep
-      [ "Ensure that you are using a valid FOSSA_API_KEY."
-      , "Refer to " <> pretty apiKeyUrl <> " for guidance on how to generate and retrieve your API key."
-      ]
+    createErrataWithHeaderOnly "Ensure that you are using a valid FOSSA_API_KEY. Refer to the provided documentation for guidance on how to generate and retrieve your API key."
+
+data TokenTypeErr = TokenTypeErr
+instance ToDiagnostic TokenTypeErr where
+  renderDiagnostic :: TokenTypeErr -> Errata
+  renderDiagnostic TokenTypeErr =
+    Errata (Just "Invalid API token type") [] $ Just "The action you are trying to perform requires a `Full Access` API token"
+
+data SubscriptionTypeErr = SubscriptionTypeErr
+instance ToDiagnostic SubscriptionTypeErr where
+  renderDiagnostic :: SubscriptionTypeErr -> Errata
+  renderDiagnostic SubscriptionTypeErr =
+    Errata (Just "Invalid subscription type") [] $ Just "The action you are trying to perform requires a premium subscription"
+
+projectPermissionErrHeader :: Text
+projectPermissionErrHeader = "Invalid project permission"
+
+data ProjectPermissionErr
+  = CreateProjectPermissionErr
+  | EditProjectPermissionErr
+  | CreateTeamProjectPermissionErr
+  | CreateProjectOnlyToTeamPermissionErr
+
+instance ToDiagnostic ProjectPermissionErr where
+  renderDiagnostic :: ProjectPermissionErr -> Errata
+  renderDiagnostic CreateProjectPermissionErr =
+    Errata (Just projectPermissionErrHeader) [] $ Just "You do not have permission to create projects for your Organization"
+  renderDiagnostic EditProjectPermissionErr =
+    Errata (Just projectPermissionErrHeader) [] $ Just "You do not have permission to edit projects for your Organization"
+  renderDiagnostic CreateTeamProjectPermissionErr =
+    Errata (Just projectPermissionErrHeader) [] $ Just "You do not have permission to create projects for the specified team"
+  renderDiagnostic CreateProjectOnlyToTeamPermissionErr =
+    Errata (Just projectPermissionErrHeader) [] $ Just "You only have permission to create projects for your team"
+
+releaseGroupPermissionErrHeader :: Text
+releaseGroupPermissionErrHeader = "Invalid release group permission"
+
+data ReleaseGroupPermissionErr
+  = EditReleaseGroupPermissionErr
+  | CreateTeamProjectsForReleaseGroupPermissionErr
+
+instance ToDiagnostic ReleaseGroupPermissionErr where
+  renderDiagnostic :: ReleaseGroupPermissionErr -> Errata
+  renderDiagnostic EditReleaseGroupPermissionErr =
+    Errata (Just releaseGroupPermissionErrHeader) [] $ Just "You do not have permission to edit the specified release group"
+  renderDiagnostic CreateTeamProjectsForReleaseGroupPermissionErr =
+    Errata (Just releaseGroupPermissionErrHeader) [] $ Just "You do not have permission to add projects to all teams belonging to the specified release group"