+++ src/Container/Docker/OciManifest.hs
@@ -25,13 +25,16 @@ import Container.Docker.SourceParser (
  )
 import Control.Effect.Diagnostics (ToDiagnostic, renderDiagnostic)
 import Data.Aeson (FromJSON (parseJSON), withObject, withText, (.:))
+import Data.Error (SourceLocation, createEmptyBlock)
 import Data.List.NonEmpty ((<|))
 import Data.List.NonEmpty qualified as NonEmpty
 import Data.String.Conversion (toString)
 import Data.Text (Text)
 import Data.Text qualified as Text
-import Effect.Logger (vsep)
-import Prettyprinter (indent, line, pretty)
+import Effect.Logger (renderIt, vsep)
+import Errata (errataSimple)
+import Errata.Types (Errata)
+import Prettyprinter (indent, line)
 
 supportedManifestKinds :: [Text]
 supportedManifestKinds =
@@ -176,17 +179,21 @@ mkLayerTarFileName (OciManifestLayer (RepoDigest digest) _) = removeDigestAlgori
 removeDigestAlgorithm :: Text -> Text
 removeDigestAlgorithm = snd . Text.breakOnEnd ":"
 
-data NotSupportedManifestFmt = NotSupportedManifestFmt Text RegistryImageSource
+data NotSupportedManifestFmt
+  = NotSupportedManifestFmt SourceLocation Text RegistryImageSource
 
 instance ToDiagnostic NotSupportedManifestFmt where
-  renderDiagnostic (NotSupportedManifestFmt fmt imgSrc) =
-    vsep
-      [ pretty $ "Manifest format is not supported: " <> fmt
-      , line <> "Workaround:" <> line
-      , indent 2 $
-          vsep
-            [ "Export the image:"
-            , line
-            , suggestDockerExport imgSrc
-            ]
-      ]
+  renderDiagnostic :: NotSupportedManifestFmt -> Errata
+  renderDiagnostic (NotSupportedManifestFmt srcLoc fmt imgSrc) = do
+    let body =
+          renderIt $
+            vsep
+              [ "Workaround:" <> line
+              , indent 2 $
+                  vsep
+                    [ "Export the image:"
+                    , line
+                    , suggestDockerExport imgSrc
+                    ]
+              ]
+    errataSimple (Just $ "Manifest format is not supported: " <> fmt) (createEmptyBlock srcLoc) (Just body)