+++ src/Strategy/Conan/ConanGraph.hs
@@ -10,7 +10,7 @@ module Strategy.Conan.ConanGraph (
   toDependency,
 ) where
 
-import Control.Effect.Diagnostics (Diagnostics, errCtx, run)
+import Control.Effect.Diagnostics (Diagnostics, errCtx, errHelp, run)
 import Data.Aeson (
   FromJSON (parseJSON),
   Key,
@@ -48,10 +48,10 @@ import Effect.Exec (
   execJson,
  )
 import Effect.Grapher (Grapher, deep, direct, edge, evalGrapher)
+import Errata (Errata (..))
 import Graphing (Graphing)
 import Network.HTTP.Types (renderQueryText)
 import Path (Abs, Dir, Path)
-import Prettyprinter (indent, vsep)
 import Strategy.Conan.Version (guardConanVersion2Gt)
 
 -- | Represents `conan install . -f json`.
@@ -302,19 +302,18 @@ analyzeFromConanGraph dir = do
   -- would ensure source code is always retrieved. Also, the
   -- equivalent command of "conan info ." in conan v1
   -- does not provide used settings for the dependency.
-  errCtx ConanV2IsRequired $ guardConanVersion2Gt dir
+  errCtx ConanV2IsRequiredCtx $ errHelp ConanV2IsRequiredHelp $ guardConanVersion2Gt dir
 
   conanGraph <- execJson dir $ conanV2GraphCmd []
   pure $ buildGraph conanGraph
 
-data ConanV2IsRequired = ConanV2IsRequired
+data ConanV2IsRequired
+  = ConanV2IsRequiredCtx
+  | ConanV2IsRequiredHelp
 instance ToDiagnostic ConanV2IsRequired where
-  renderDiagnostic (ConanV2IsRequired) =
-    vsep
-      [ "Conan analysis requires conan v2.0.0 or greater"
-      , ""
-      , indent 2 $
-          vsep
-            [ "Ensure you are using conan v2. You can check this by running, conan --version"
-            ]
-      ]
+  renderDiagnostic ConanV2IsRequiredCtx = do
+    let header = "Conan analysis requires conan v2.0.0 or greater"
+    Errata (Just header) [] Nothing
+  renderDiagnostic ConanV2IsRequiredHelp = do
+    let header = "Ensure you are using conan v2 by running, conan --version"
+    Errata (Just header) [] Nothing