+++ src/Strategy/Python/ReqTxt.hs
@@ -6,12 +6,13 @@ module Strategy.Python.ReqTxt (
 import Control.Effect.Diagnostics
 import Control.Monad (void)
 import Data.Foldable (asum)
+import Data.String.Conversion (toText)
 import Data.Text (Text)
 import Data.Void (Void)
 import Effect.ReadFS
+import Errata (Errata (..))
 import Graphing (Graphing)
 import Path
-import Prettyprinter (Pretty (pretty), vsep)
 import Strategy.Python.Pip (PythonPackage)
 import Strategy.Python.Util
 import Text.Megaparsec
@@ -20,19 +21,19 @@ import Types
 
 analyze' :: (Has ReadFS sig m, Has Diagnostics sig m) => Maybe [PythonPackage] -> Path Abs File -> m (Graphing Dependency)
 analyze' packages file = do
-  reqs <- errCtx (ReqsTxtFailed file) $ readContentsParser requirementsTxtParser file
+  reqs <- errCtx (ReqsTxtFailedCtx file) $ errHelp ReqsTxtFailedHelp $ readContentsParser requirementsTxtParser file
   context "Building dependency graph" $ pure (buildGraph packages reqs)
 
-newtype ReqsTxtFailed = ReqsTxtFailed (Path Abs File)
+data ReqsTxtFailed
+  = ReqsTxtFailedCtx (Path Abs File)
+  | ReqsTxtFailedHelp
 instance ToDiagnostic ReqsTxtFailed where
-  renderDiagnostic (ReqsTxtFailed path) =
-    vsep
-      [ pretty $ "Failed to parse: " <> show path
-      , ""
-      , "We occasionally find files we think are python requirements.txt files but"
-      , "aren't. If this file isn't a python requirements.txt file, this error can"
-      , "be safely ignored."
-      ]
+  renderDiagnostic (ReqsTxtFailedCtx path) = do
+    let header = "Failed to parse: " <> toText (show path)
+    Errata (Just header) [] Nothing
+  renderDiagnostic ReqsTxtFailedHelp = do
+    let header = "Ignore this error if this file isn't a python requirements.txt file."
+    Errata (Just header) [] Nothing
 
 type Parser = Parsec Void Text