+++ src/Strategy/Swift/PackageSwift.hs
@@ -13,7 +13,7 @@ module Strategy.Swift.PackageSwift (
 ) where
 
 import Control.Applicative (Alternative ((<|>)), optional)
-import Control.Effect.Diagnostics (Diagnostics, context, errCtx, fatalText, recover, warnOnErr)
+import Control.Effect.Diagnostics (Diagnostics, context, errCtx, errDoc, errHelp, fatalText, recover, warnOnErr)
 import Control.Monad (void)
 import Data.Foldable (asum)
 import Data.Map.Strict qualified as Map
@@ -25,7 +25,7 @@ import Diag.Common (MissingDeepDeps (MissingDeepDeps))
 import Effect.ReadFS (Has, ReadFS, readContentsJson, readContentsParser)
 import Graphing (Graphing, deeps, directs, induceJust, promoteToDirect)
 import Path
-import Strategy.Swift.Errors (MissingPackageResolvedFile (..))
+import Strategy.Swift.Errors (MissingPackageResolvedFile (..), MissingPackageResolvedFileHelp (..), swiftFossaDocUrl, swiftPackageResolvedRef, xcodeCoordinatePkgVersion)
 import Strategy.Swift.PackageResolved (SwiftPackageResolvedFile, resolvedDependenciesOf)
 import Text.Megaparsec (
   MonadParsec (takeWhile1P, try),
@@ -217,6 +217,10 @@ analyzePackageSwift manifestFile resolvedFile = do
         recover
           . warnOnErr MissingDeepDeps
           . errCtx (MissingPackageResolvedFile manifestFile)
+          . errHelp MissingPackageResolvedFileHelp
+          . errDoc swiftFossaDocUrl
+          . errDoc swiftPackageResolvedRef
+          . errDoc xcodeCoordinatePkgVersion
           $ fatalText "Package.resolved file was not discovered"
     Just packageResolved -> context "Identifying dependencies in Package.resolved" $ readContentsJson packageResolved