+++ test/Test/MockApi.hs
@@ -236,6 +236,7 @@ matchExpectation a@(UploadContentForReachability{}) (ApiExpectation _ requestExp
 matchExpectation a@(UploadBuildForReachability{}) (ApiExpectation _ requestExpectation b@(UploadBuildForReachability{}) resp) = checkResult requestExpectation a b resp
 matchExpectation a@(GetTokenType{}) (ApiExpectation _ requestExpectation b@(GetTokenType{}) resp) = checkResult requestExpectation a b resp
 matchExpectation a@(GetSubscription{}) (ApiExpectation _ requestExpectation b@(GetSubscription{}) resp) = checkResult requestExpectation a b resp
+matchExpectation a@(GetCustomBuildPermissons{}) (ApiExpectation _ requestExpectation b@(GetCustomBuildPermissons{}) resp) = checkResult requestExpectation a b resp
 matchExpectation _ _ = Nothing
 
 -- | Handles a request in the context of the mock API.